Split into library and driver
* Updated CMakeLists.txt to make libfractal.a
* Created driver main.c that uses the library
